Update remember_me.rst
If your session is removed or deleted from /var/lib/php/sessions but you have a valid remember be cookie, the user not longer has the role `IS_AUTHENTICATED_FULLY` but does have the role `IS_AUTHENTICATED_REMEMBERED`, couldn't see this in the documentation currently. So belive this is a nice too have. May also be related to this ticket I found.

https://github.com/symfony/symfony/issues/9203